cmocka: Add support for expect_(not_)double(_count)()

+++ tests/test_expect_in_range.c
@@ -10,8 +10,7 @@
 #include <stdio.h>
 #include <string.h>
 
-
-static void mock_test_float(double value)
+static void mock_test_float(float value)
 {
     check_expected_float(value);
 }
